requirements.txt
langgraph>=0.0.60
langchain_openai>=0.0.8
python-dotenv>=1.0.0
bedrock_agentcore